rabbit_api/sender/requirements.txt
fastapi==0.85.1
orjson==3.8.0
pydantic==1.10.2
uvicorn==0.19.0
gunicorn==20.1.0
SQLAlchemy==1.4.42
pika==1.3.1
alembic==1.7.1
asyncpg==0.24.0
pytest==7.2.0
requests==2.28.1
python-dotenv==0.19.2
Jinja2==3.1.2
psycopg2-binary==2.9.4
psycopg2==2.9.6